requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.9.0
bootstrap4==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==2.0.13
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==2.2.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
